requirements.txt
apturl==0.5.2
avro-python3==1.10.2
bcrypt==3.2.0
blinker==1.4
brainflow==5.8.1
Brlapi==0.8.3
cachetools==5.5.0
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.6
colorlog==6.7.0
confluent-kafka==2.3.0
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
Cython==3.0.4
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
filelock==3.16.1
fonttools==4.40.0
-e git+https://github.com/twabi/frameMQ.git@7315c33af20113067930a37f5933a54a33c106ea#egg=FrameMQ
future==0.18.2
gcloud==0.17.0
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
importlib-metadata==4.6.4
importlib-resources==5.12.0
iniconfig==2.0.0
jeepney==0.7.1
Jinja2==3.1.2
jws==0.1.3
kafka-python==2.0.2
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lz4==4.3.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.7.1
mne==1.4.2
monotonic==1.6
more-itertools==8.10.0
mozjpeg-lossless-optimization==1.1.3
netifaces==0.11.0
neurosity==1.0.2
nptyping==1.4.4
numpy==1.24.3
oauth2client==3.0.0
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paho-mqtt==1.6.1
pandas==2.0.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.5.0
ping3==4.0.8
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.7.0
promise==2.3
protobuf==4.23.3
psutil==6.0.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.4.3
pycryptodomex==3.18.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyGObject-stubs==2.12.0
pyguetzli==1.0.15
PyJWT==2.3.0
pylsl==1.16.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyproject-api==1.8.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqtgraph==0.12.4
Pyrebase==3.0.27
pyRFC3339==1.1
pytest==8.3.4
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-jwt==2.0.1
PyTurboJPEG==1.7.2
pytz==2023.3
pyxdg==0.27
PyYAML==6.0
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.7.0
rsa==4.9
scipy==1.10.1
SecretStorage==3.3.1
sinetstream==1.8.1
sinetstream-comp-lz4==1.8.1
sinetstream-iostream==1.8.1
sinetstream-kafka==1.8.0
sinetstream-mqtt==1.8.0
sinetstream-type-image==1.8.1
six==1.16.0
systemd-python==234
tomli==2.2.1
tox==4.23.2
tqdm==4.65.0
typing_extensions==4.12.2
typish==1.9.3
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.3
usb-creator==0.3.7
uuid==1.30
vboxapi==1.0
vidgear==0.3.2
virtualenv==20.28.0
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==3.15.0
zstandard==0.21.0
